+++ CHANGELOG.md
@@ -18,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `unionMesh()` to `PGS_ShapeBoolean`. Quickly and efficiently unions/merges the faces of a mesh-like shape together.
 * `setAllStrokeToFillColor()` to `PGS_Conversion`. Sets the stroke color to the fill color for a PShape and all its descendants (separately).
 * A number of new primitives to `PGS_Construction`: `serpinskiCurve()`, `spiral()`, `capsule()`.
-* `quadrangulation()` to `PGS_Meshing`. Produces a quadrangulation from a triangulation via triangle subdivision.
+* `splitQuadrangulation()` and `spiralQuadrangulation()` to `PGS_Meshing`. These methods produces a quadrangulation from a triangulation / point set using differing algorithms.
 * `dualFaces()` to `PGS_Meshing`. Generates a mesh-like shape consisting of polygonal faces of a dual graph of a given triangulation.
 
 ### Changed